Fix propagation of cookie banner between pages
This applies the change from https://github.com/inventables/easel/pull/10069 to set the cookie banner cookie's `path` to allow it to be read from any paths on our domain—by default cookies are scoped to the beginning part of the path of the page where they're set (e.g. `/products` when set from a product page). Also sets `max-age` to 365 days so that the cookie outlasts the session.